Changing explainability structure to a more scalable one.
