+++ fedot/core/explainability/explainer.py
@@ -1,0 +1,17 @@
+from abc import abstractmethod
+
+
+class Explainer:
+    """
+    An abstract class for various explanation methods.
+    """
+    def __init__(self, model):
+        self.model = model
+
+    @abstractmethod
+    def explain(self, *args, **kwargs):
+        raise NotImplementedError
+
+    @abstractmethod
+    def output(self, *args, **kwargs):
+        raise NotImplementedError
